aws-java/RDS/rds-demos/rds-with-aws-secretmanager-jdbc/rds-aws-secretmanager-jdbc-demo/src/main/java/com/myapps/rds/controller/EmployeeController.java
package com.myapps.rds.controller;

import com.myapps.rds.model.Employee;
import com.myapps.rds.service.EmployeeService;
import lombok.RequiredArgsConstructor;
import org.springframework.web.bind.annotation.*;

import java.util.List;

@RestController
@RequiredArgsConstructor
@RequestMapping("/api/v1/employees")
public class EmployeeController {

    private final EmployeeService employeeService;


    @PostMapping
    public Employee createEmployee(@RequestBody Employee employee) {
        return employeeService.createEmployee(employee);
    }

    @GetMapping
    public List<Employee> getAll() {
        return employeeService.getAll();
    }

}
